requirements.txt
openai
pyautogui
pygetwindow
pyperclip
pillow
opencv-python
psutil
python-docx
openpyxl
python-pptx
pywin32
pyyaml
vosk
sounddevice
numpy
